fix: potentialWinnings 필드 DECIMAL(20,2)로 수정 (멀티 배팅 소수점 오류 해결)

+++ server/models/betModel.js
@@ -31,7 +31,7 @@ const Bet = sequelize.define('Bet', {
     allowNull: false
   },
   potentialWinnings: {
-    type: DataTypes.INTEGER,
+    type: DataTypes.DECIMAL(20, 2),
     allowNull: false
   },
   status: {
